pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-bot"
version = "0.1.0"
description = "Telegram bot for cloning and synthesizing speech"
readme = "README.md"
authors = [{ name = "Speech Bot Team" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "aiogram==3.*",
    "python-dotenv",
    "aiosqlite",
    "TTS",
    "pydub",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.3.0",
    "pytest>=8.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]
extend-exclude = "bot\\.db"

[tool.ruff]
target-version = "py310"
line-length = 88
extend-exclude = ["bot.db"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["speech"]
combine-as-imports = true

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

